Update drillhole.py
Address #117
geoh5py.objects.drillhole.Drillhole.locations dtype inconsistency
Changed dtype of locations

+++ geoh5py/objects/drillhole.py
@@ -190,9 +190,9 @@ class Drillhole(Points):
         ):
             lengths = self.surveys["Depth"][1:] - self.surveys["Depth"][:-1]
             self._locations = np.c_[
-                self.collar["x"] + np.cumsum(np.r_[0.0, lengths * self.deviation_x]),
-                self.collar["y"] + np.cumsum(np.r_[0.0, lengths * self.deviation_y]),
-                self.collar["z"] + np.cumsum(np.r_[0.0, lengths * self.deviation_z]),
+                self.collar["x"] + np.cumsum(np.r_[0.0, lengths * self.deviation_x]).astype('float64'),
+                self.collar["y"] + np.cumsum(np.r_[0.0, lengths * self.deviation_y]).astype('float64'),
+                self.collar["z"] + np.cumsum(np.r_[0.0, lengths * self.deviation_z]).astype('float64'),
             ]
 
         return self._locations
